fix: Improve react-native calls
Same as Dialog component, could be refactored into single hook
With cozy-bar context, cssVar are unreliable.
It is safer to rely on already used values in the DOM for now
We don't want to modify top and bottom color in this scenario
