+++ react/Modal/index.jsx
@@ -14,6 +14,8 @@ import ModalFooter from './ModalFooter'
 import ModalButtons from './ModalButtons'
 import AnimatedContentHeader from './AnimatedContentHeader'
 import ModalBackButton from './ModalBackButton'
+import { ModalEffects } from './ModalEffects'
+import { isFlagshipApp } from 'cozy-device-helper'
 
 const ModalDescription = ModalContent
 
@@ -94,6 +96,7 @@ class Modal extends Component {
 
     return (
       <Portal into={into}>
+        {isFlagshipApp() && <ModalEffects />}
         <div className={cx(styles['c-modal-container'], containerClassName)}>
           <Overlay
             onEscape={closable ? dismissAction : undefined}
